Create Main
Thank you to my many, many, many fans!

Rewrote the whole thing from scratch using Classes and improved resolve() techniques. Down to ~1,100 lines from over 5,000!

Improved legibility with colors from colorama and unicode card symbols, improved game flow and information delivery, added option to peek at cards (rather than forcing a peek every turn) so you don't have to close your eyes as often.

Barring any remaining bugs in the resolve() function, all that would be left is to optimize it so it doesn't need to run as many calculations. Probably the slowest part is quadcheck(), which uses 4 for loops to check 7 cards for matches, generating 7^4 (2,401) operations, many of which are totally unecessary (e.g. checking if cards with the same name have the same name hundreds of times). Moreover, things like tripcheck() and paircheck() can be lumped in to the quadcheck() function (if it finds a trip while it's looking for quads, it ought to remember that!). Likewise for the strflshcheck() with respect to flushcheck() and straightcheck().