Create README.md
Recently updated as of 2/13/2025 with major improvements both in gameplay and coding elegance.
It's all in just one file, Main.py.
Requires colorama. If for whatever reason you don't want to install that, have fun reformatting all the card names.
For some reason, the colorama BLUE is much darker in CMD than in Pycharm. Dunno what's up with that, but you should probably just run it in your IDE.
Note the recommended terminal font size of 30 pt. Increasing it from the default will great improve legibility of the cards.